+++ lib/src/classes/events.dart
@@ -1,0 +1,9 @@
+// ignore_for_file: constant_identifier_names
+
+class Events {
+  static const Ready = "READY";
+  static const GuildCreate = "GUILD_CREATE";
+  static const GuildDelete = "GUILD_DELETE";
+  static const InteractionCreate = "INTERACTION_CREATE";
+  static const MessageCreate = "MESSAGE_CREATE";
+}