+++ lib/src/classes/guild/unavailable_guild.dart
@@ -3,9 +3,9 @@ import "guild.dart";
 class UnavailableGuild {
   late String id;
   late bool unavailable;
-  late Guild? notUpdatedData;
+  late Guild? notUpdatedGuild;
 
-  UnavailableGuild(this.id, { this.notUpdatedData }) {
+  UnavailableGuild(this.id, { this.notUpdatedGuild }) {
     unavailable = true;
   }
 }