+++ lib/src/classes/message/message.dart
@@ -1,12 +1,27 @@
+import "embed.dart";
+
 /// Represents a message on Discord.
 class Message {
   String? content;
-  Message({ this.content });
+  List<Embed>? embeds;
+
+  Message({ this.content, this.embeds });
 
   /// Returns an object rapresentation of the message
   Map<String, dynamic> exportable() {
-    return {
-      "content": content,
-    };
+    Map<String, dynamic> a = {};
+    List<Map<String, dynamic>> e = [];
+    if (embeds != null && embeds!.isNotEmpty) {
+      for (var embed in embeds!) {
+        e.add(embed.exportable());
+      }
+    }
+    if (e.isNotEmpty) {
+      a["embeds"] = e;
+    }
+    if (content != null) {
+      a["content"] = content;
+    }
+    return a;
   }
 }