+++ lib/src/classes/message/message_sent.dart
@@ -2,8 +2,9 @@ import 'message.dart';
 
 /// Represents a sent message on Discord.
 class MessageSent extends Message {
-  String id = '';
-  MessageSent(Message msg, { required this.id }) {
-    content = msg.content;
-  }
+  late String id;
+  late String channelID;
+  late String authorID;
+
+  MessageSent(Message msg, this.authorID, this.id, this.channelID): super(content: msg.content, embeds: msg.embeds);
 }